LPS-157775 extract methods of upgrade dl configurations processes to reuse several times
