+++ modules/apps/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/upgrade/v3_2_4/DLSizeLimitConfigurationUpgradeProcess.java
@@ -5,16 +5,8 @@
 
 package com.liferay.document.library.internal.upgrade.v3_2_4;
 
-import com.liferay.document.library.configuration.DLConfiguration;
-import com.liferay.document.library.internal.configuration.DLSizeLimitConfiguration;
-import com.liferay.petra.string.StringPool;
+import com.liferay.document.library.internal.upgrade.helper.DLConfigurationUpgradeHelper;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
-
-import java.util.Dictionary;
-
-import org.osgi.service.cm.Configuration;
-import org.osgi.service.cm.ConfigurationAdmin;
 
 /**
  * @author Adolfo Pérez
@@ -22,59 +14,16 @@ import org.osgi.service.cm.ConfigurationAdmin;
 public class DLSizeLimitConfigurationUpgradeProcess extends UpgradeProcess {
 
 	public DLSizeLimitConfigurationUpgradeProcess(
-		ConfigurationAdmin configurationAdmin) {
+		DLConfigurationUpgradeHelper dlConfigurationUpgradeHelper) {
 
-		_configurationAdmin = configurationAdmin;
+		_dlConfigurationUpgradeHelper = dlConfigurationUpgradeHelper;
 	}
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		long fileMaxSize = _updateDLConfiguration();
-
-		if (fileMaxSize > 0) {
-			_updateDLSizeLimitConfiguration(fileMaxSize);
-		}
+		_dlConfigurationUpgradeHelper.updateDLSizeLimitConfiguration();
 	}
 
-	private long _updateDLConfiguration() throws Exception {
-		Configuration[] configurations = _configurationAdmin.listConfigurations(
-			"(service.pid=" + DLConfiguration.class.getName() + ")");
-
-		if (configurations == null) {
-			return 0;
-		}
-
-		Configuration configuration = configurations[0];
-
-		Dictionary<String, Object> dictionary = configuration.getProperties();
-
-		if (dictionary == null) {
-			return 0;
-		}
-
-		Long fileMaxSize = (Long)dictionary.remove("fileMaxSize");
-
-		if (fileMaxSize == null) {
-			return 0;
-		}
-
-		configuration.update(dictionary);
-
-		return fileMaxSize;
-	}
-
-	private void _updateDLSizeLimitConfiguration(long fileMaxSize)
-		throws Exception {
-
-		Configuration configuration = _configurationAdmin.getConfiguration(
-			DLSizeLimitConfiguration.class.getName(), StringPool.QUESTION);
-
-		configuration.update(
-			HashMapDictionaryBuilder.put(
-				"fileMaxSize", fileMaxSize
-			).build());
-	}
-
-	private final ConfigurationAdmin _configurationAdmin;
+	private final DLConfigurationUpgradeHelper _dlConfigurationUpgradeHelper;
 
 }
